requirements.txt
asgiref==3.10.0
asttokens==3.0.0
bleach==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.4
decorator==5.2.1
Django==5.1.4
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
Faker==37.11.0
idna==3.11
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kaggle==1.7.4.5
matplotlib-inline==0.1.7
numpy==2.3.4
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
